legacy/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
dj-database-url==2.3.0
Django==5.1.1
django-allauth==65.0.2
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-mdeditor==0.1.20
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
mistune==3.1.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
Pygments==2.19.1
pytesseract==0.3.13
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.8.2
